BE cleanup

+++ Backend/src/VidiVideo.Application/Videos/DeleteVideoCommandHandler.cs
@@ -1,6 +1,8 @@
-﻿using VidiVideo.Application.Abstractions.Repositories;
+﻿using VidiVideo.Application.Abstractions;
+using VidiVideo.Application.Abstractions.Repositories;
 using VidiVideo.Application.Common;
 using VidiVideo.Application.Exceptions;
+using VidiVideo.Domain.Constants;
 
 namespace VidiVideo.Application.Videos
 {
@@ -8,16 +10,25 @@ namespace VidiVideo.Application.Videos
     {
         private readonly IVideoRepository _repo;
         private readonly IUnitOfWork _unitOfWork;
+        private readonly ICurrentUser _currentUser;
 
-        public DeleteVideoCommandHandler(IVideoRepository repo, IUnitOfWork unitOfWork)
+        public DeleteVideoCommandHandler(IVideoRepository repo, IUnitOfWork unitOfWork, ICurrentUser currentUser)
         {
             _repo = repo;
             _unitOfWork = unitOfWork;
+            _currentUser = currentUser;
         }
 
         public async Task<bool> HandleAsync(DeleteVideoCommand command, CancellationToken cancellationToken)
         {
-            var video = await _repo.GetVideoByIdAsync(command.Id) ?? throw new NotFoundException("Video doesn't exist");
+            if (!await _repo.ExistsByIdAsync(command.Id))
+                throw new NotFoundException("Video doesn't exist");
+
+            var creatorId = _currentUser.UserId ?? throw new UnauthorizedException("Must be logged in");
+
+            if (!_currentUser.IsInRole(AppRoles.Admin))
+                if (!await _repo.CheckOwnershipAsync(creatorId, command.Id))
+                    throw new UnauthorizedException("You are not the owner of this video");
 
             await _repo.DeleteVideoAsync(command.Id);
